Hand-merged stuff added to BTC-node-service since merge with monorepo
Merged: 9372e7e2ddf8..9f25a62e677e
That is commits:
 * cf01b5762c1e logs refactor
 * 4afb7f46fc26 Mempool schema fix

+++ ns-btc/utils.go
@@ -776,7 +776,7 @@ func delSpOutToGenerated(del store.DeleteSpendableOutput) pb.ReqDeleteSpOut {
 
 func newMempoolRecord(category int, hashTX string) store.MempoolRecord {
 	return store.MempoolRecord{
-		Category: category,
+		Category: int64(category),
 		HashTX:   hashTX,
 	}
 }
